Change the IV to be a constant nonce (at it was originally).
Somehow over time the IV changed meaning and happened to
take the IV from the generated bytes (like CBC), however
it doesn't provide anything anyway, live the IV to be
a constant that is xored to the counter.

Only having the counter is enough already, maybe later version
will remove this altogether.